Add files via upload
This is a java console app that has three functions. The first function asks the user to input a word and then reverses that word. The second function checks to see if a word is a palindrome. The last function takes two user inputs of integers and choose a random number between the two numbers